+++ test/client.ts
@@ -31,6 +31,10 @@ import {
   HTTPNetworkInterface,
 } from '../src/networkInterface';
 
+import { getQueryDefinition } from '../src/queries/getFromAST';
+
+import { addTypenameToSelectionSet } from '../src/queries/queryTransform';
+
 import mockNetworkInterface from './mocks/mockNetworkInterface';
 
 import * as chaiAsPromised from 'chai-as-promised';
@@ -298,7 +302,7 @@ describe('client', () => {
       apollo: {
         queries: {
           '0': {
-            queryString: print(query),
+            queryString: print(getQueryDefinition(query)),
             query: {
               id: 'ROOT_QUERY',
               typeName: 'Query',
@@ -461,6 +465,112 @@ describe('client', () => {
         done();
       },
     });
+  });
+
+  it('should be able to transform queries', (done) => {
+    const query = gql`
+      query {
+        author {
+          firstName
+          lastName
+        }
+      }`;
+    const transformedQuery = gql`
+      query {
+        author {
+          firstName
+          lastName
+          __typename
+        }
+        __typename
+      }`;
+
+    const result = {
+      'author': {
+        'firstName': 'John',
+        'lastName': 'Smith',
+      },
+    };
+    const transformedResult = {
+      'author': {
+        'firstName': 'John',
+        'lastName': 'Smith',
+        '__typename': 'Author',
+      },
+      '__typename': 'RootQuery',
+    };
+
+    const networkInterface = mockNetworkInterface(
+    {
+      request: { query },
+      result: { data: result },
+    },
+    {
+      request: { query: transformedQuery },
+      result: { data: transformedResult },
+    });
+
+    const client = new ApolloClient({
+      networkInterface,
+      queryTransformer: addTypenameToSelectionSet,
+    });
+
+    client.query({ query }).then((actualResult) => {
+      assert.deepEqual(actualResult.data, transformedResult);
+      done();
+    });
+  });
+
+  it('should be able to transform queries on forced fetches', (done) => {
+    const query = gql`
+      query {
+        author {
+          firstName
+          lastName
+        }
+      }`;
+    const transformedQuery = gql`
+      query {
+        author {
+          firstName
+          lastName
+          __typename
+        }
+        __typename
+      }`;
+    const result = {
+      'author': {
+        'firstName': 'John',
+        'lastName': 'Smith',
+      },
+    };
+    const transformedResult = {
+      'author': {
+        'firstName': 'John',
+        'lastName': 'Smith',
+        '__typename': 'Author',
+      },
+      '__typename': 'RootQuery',
+    };
+    const networkInterface = mockNetworkInterface(
+    {
+      request: { query },
+      result: { data: result },
+    },
+    {
+      request: { query: transformedQuery },
+      result: { data: transformedResult },
+    });
+
+    const client = new ApolloClient({
+      networkInterface,
+      queryTransformer: addTypenameToSelectionSet,
+    });
+    client.query({ forceFetch: true, query }).then((actualResult) => {
+      assert.deepEqual(actualResult.data, transformedResult);
+      done();
+    });
+
   });
 
   describe('accepts dataIdFromObject option', () => {